Split marshaler implementation into smaller files
